+++ packages/utils/src/index.ts
@@ -78,24 +78,3 @@ export * from "./url";
 export * from "./windows-acl";
 export * from "./which";
 export * from "./yaml-sync";
-
-function isPlainObject(val: object): val is Record<string, unknown> {
-	return Object.getPrototypeOf(val) === Object.prototype || Array.isArray(val);
-}
-
-export function structuredCloneJSON<T>(value: T): T {
-	// primitives|null|undefined, copy
-	if (!value || typeof value !== "object") {
-		return value;
-	}
-
-	// deep clone
-	if (isPlainObject(value)) {
-		try {
-			return structuredClone(value);
-		} catch {
-			// might still fail due to nested structures
-		}
-	}
-	return JSON.parse(JSON.stringify(value)) as T;
-}